backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
psycopg2-binary==2.9.9
redis==5.0.4
aioredis==2.0.1
python-multipart==0.0.9
pydantic-settings==2.2.1
websockets==12.0
httpx==0.27.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
